{fix} now strlen does return a t_size value and not an integer anymore

+++ functions/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: ldevoude <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/13 15:10:20 by ldevoude          #+#    #+#             */
-/*   Updated: 2024/11/14 13:13:04 by ldevoude         ###   ########.fr       */
+/*   Updated: 2024/11/14 13:45:12 by ldevoude         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,24 +14,24 @@
 //#include <ctype.h>
 //#include <string.h>
 
-int	ft_strlen(char str[]);
+size_t	ft_strlen(char str[]);
 /*
 int	main(void)
 {
 	char	str[42] = "forty-two";
-	printf("origin : %lu\n ft : %i\n",strlen(str), ft_strlen(str));
+	printf("origin : %zu\n ft : %zu\n",strlen(str), ft_strlen(str));
 
 	return(0);
 }
 */
 
-int	ft_strlen(char str[])
+size_t	ft_strlen(char str[])
 {
-	int	i;
-	
+	size_t	i;
+
 	i = 0;
-	while (str[i] != '\0') 
-	{ 
+	while (str[i] != '\0')
+	{
 		i++;
 	}
 	return (i);
